requirements.txt
sounddevice
numpy
scipy
openwakeword
onnxruntime
ollama
tavily-python
Pillow
